requirements.txt
Django==3.2.9
django-environ==0.8.1
psycopg2-binary==2.9.2
djangorestframework==3.12.4